# backend/openapi.yaml
openapi: 3.1.0
info:
  title: GdzieZamieszkac API
  version: 1.0.0-draft
  summary: Read-only district data for one city per deployment.
  description: |
    Serves the audited district data of one city (Warsaw or Krakow, chosen by the `CITY` setting of the deployment). There is no city
    parameter and no `/cities` endpoint: each deployment is a separate product.

    Rules that hold for every endpoint (see `docs/BACKEND_PLAN.md` and `docs/API_CONTRACT_DRAFT.md`):
    - **Read-only, no accounts, no personal data.** The only request-time computation is `/recommend` (and the small stateless
      calculation in `rent-vs-buy`); everything else is a stored value.
    - **Nothing is translated or formatted live.** Display strings, labels, descriptions, notes and reports are stored in English and Polish.
    - **Provenance travels with every value:** data kind (`observed`, `estimated`, `proxy`), source, licence, attribution, as-of date,
      method and caveat.
    - **Honest gaps:** a metric without data in this city is reported as unavailable with a reason, never as zero or an empty value.
    - **Scores compare districts of one city only.**
    - **Stale data is served with a warning:** if the data audit failed or reports are stale, responses carry the `X-Data-Warning`
      header and `/meta` sets `stale.is_stale`; the service does not refuse to start.

...
    this version: it ships only after a backtest passes (`docs/BACKEND_PLAN.md` section 5) and will be added as a new path.
...
servers:
...
paths:
...
  /meta:
    get:
      tags: [system]
      operationId: getMeta
      summary: City, languages, data version, freshness and the attribution lines for the footer.
      parameters:
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Metadata of this deployment.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meta" }
        default:
          $ref: "#/components/responses/Problem"
  /districts:
    get:
      tags: [districts]
      operationId: listDistricts
      summary: All districts with the default livability score and two highlight values.
      parameters:
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Districts of the city (18 for Warsaw and for Krakow).
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
            X-Data-Warning: { $ref: "#/components/headers/DataWarning" }
          content:
            application/json:
              schema:
                type: object
                required: [lang, districts, score_note]
                properties:
                  lang: { $ref: "#/components/schemas/Lang" }
                  score_note: { type: string, description: 'Always says that scores compare districts of this city only.' }
                  districts:
                    type: array
                    items: { $ref: "#/components/schemas/DistrictListItem" }
        default:
          $ref: "#/components/responses/Problem"
  /districts.geojson:
    get:
      tags: [districts]
      operationId: getDistrictBoundaries
      summary: Simplified district boundaries for the map.
      description: One simplified polygon per district, with the district `code` in `properties`. Cached; changes only when boundaries are reloaded.
      responses:
        "200":
          description: A GeoJSON FeatureCollection.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
          content:
            application/geo+json:
              schema: { $ref: "#/components/schemas/BoundaryCollection" }
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}:
    get:
      tags: [districts]
      operationId: getDistrict
      summary: Everything about one district, grouped by category, with the rank of each metric within the city.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: The district.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
            X-Data-Warning: { $ref: "#/components/headers/DataWarning" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictDetail" }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}/report:
    get:
      tags: [districts]
      operationId: getDistrictReport
      summary: The stored area report and the default livability score.
      description: The report is generated at ingestion and stored in both languages. If the Polish text is missing, English is returned and `lang_fallback` is true.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: The report.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-Data-Warning: { $ref: "#/components/headers/DataWarning" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}/similar:
    get:
      tags: [districts]
      operationId: getSimilarDistricts
      summary: Districts of the same city with the most similar profile.
      description: Cosine similarity of the percentile vectors of the scored metrics (the ones the livability score uses). Neutral metrics are not used.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/Lang"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 10, default: 3 }
      responses:
        "200":
          description: Similar districts, most similar first.
          content:
            application/json:
              schema:
                type: object
                required: [district, lang, similar, method]
                properties:
                  district: { $ref: "#/components/schemas/DistrictCode" }
                  lang: { $ref: "#/components/schemas/Lang" }
                  method: { type: string }
                  similar:
                    type: array
                    items:
                      type: object
                      required: [code, name, similarity]
                      properties:
                        code: { $ref: "#/components/schemas/DistrictCode" }
                        name: { type: string }
                        similarity: { type: number, minimum: 0, maximum: 1 }
                        closest_on: { type: array, items: { type: string }, description: 'Metric keys where the two districts are closest.' }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}/series/{key}:
    get:
      tags: [districts]
      operationId: getDistrictSeries
      summary: Quarterly history of one metric for a district (price charts).
      description: |
        Only metrics with a stored history are available (the first is `sale_price_median_m2`, quarterly from 2021 in Kraków and from 2022Q4 in Warsaw, because earlier
        register records are incomplete). Quarters with fewer deeds than `min_obs` are returned with `low_confidence: true`, never dropped silently.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/MetricKey"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: The series.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Series" }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}/outlook:
    get:
      tags: [districts]
      operationId: getDistrictOutlook
      summary: Price momentum and historical range for a district. No forecast.
      description: |
        Momentum is read from the stored quarterly district history (`/series`): the change over the last four quarters and since the first stored
        quarter. The historical range is the spread of past one- and two-year price changes in the whole city (NBP data), shown as history.
        **No forward scenario is published:** the backtest in `scenario.backtest` shows that no tested method beat the naive baselines, so
        `scenario.published` is false. If a later method passes, `published` becomes true and `scenario.ranges` is filled. Never a single number,
        never a ranking of districts by predicted growth, not financial advice.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: The outlook.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Outlook" }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /districts/{code}/rent-vs-buy:
    get:
      tags: [tools]
      operationId: getRentVsBuy
      summary: Gross yield and payback for a flat of a given size (stateless calculation).
      description: |
        Uses the stored district medians: price = sale price per m2 x area, rent = rent per m2 x area, yield = rent x 12 / price,
        payback = price / (rent x 12). Asking rent over transaction price; no mortgage, tax, costs or price change. Values are `estimated`.
      parameters:
        - $ref: "#/components/parameters/DistrictCode"
        - $ref: "#/components/parameters/Lang"
        - name: area_m2
          in: query
          required: true
          schema: { type: number, minimum: 15, maximum: 250 }
      responses:
        "200":
          description: The comparison.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RentVsBuy" }
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /compare:
    get:
      tags: [tools]
      operationId: compareDistricts
      summary: Side-by-side values of two to four districts.
      parameters:
        - name: codes
          in: query
          required: true
          description: Comma-separated district codes, two to four.
          schema: { type: string, examples: ["srodmiescie,mokotow,wola"] }
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: The chosen districts in the request order, same metric shape as the district detail.
          content:
            application/json:
              schema:
                type: object
                required: [lang, districts]
                properties:
                  lang: { $ref: "#/components/schemas/Lang" }
                  districts:
                    type: array
                    minItems: 2
                    maxItems: 4
                    items: { $ref: "#/components/schemas/DistrictDetail" }
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /commute:
    get:
      tags: [tools]
      operationId: getCommute
      summary: Public-transport time from one district to every district.
      description: |
        Read from a precomputed table: simple routing over the city's GTFS, one representative weekday morning, district centroids, walking to
        and from stops. Values are `estimated` and describe typical access between areas, not a journey plan.
      parameters:
        - name: from
          in: query
          required: true
          schema: { $ref: "#/components/schemas/DistrictCode" }
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Minutes to each destination district.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Commute" }
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /personas:
    get:
      tags: [tools]
      operationId: listPersonas
      summary: Named weight presets for the recommender.
      parameters:
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Presets, stored as data. Send `weights` to `/recommend` as they are, or let the user adjust them.
          content:
            application/json:
              schema:
                type: object
                required: [lang, personas]
                properties:
                  lang: { $ref: "#/components/schemas/Lang" }
                  personas:
                    type: array
                    items: { $ref: "#/components/schemas/Persona" }
        default:
          $ref: "#/components/responses/Problem"
  /recommend:
    post:
      tags: [tools]
      operationId: recommend
      summary: Rank the districts for the user's weights and explain the ranking.
      description: |
        Percentile ranks within the city, 100 = best, weighted as requested and renormalised. Without weights the result equals the stored
        `livability_score_default`. Neutral metrics and composite indices are never inputs. Metrics without data in this city are ignored and
        listed in `missing_metrics`.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecommendRequest" }
            examples:
              balanced:
                summary: Transport and cost first
                value:
                  weights:
                    category: { transport: 3, cost: 3, safety: 2, environment: 1, amenities: 1, livability: 1 }
      responses:
        "200":
          description: Ranked districts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendResponse" }
        "422":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
    Lang:
      name: lang
      in: query
      description: Response language. If absent, the `Accept-Language` header is used, then English. Every response echoes the language used in `lang`.
      schema: { $ref: "#/components/schemas/Lang" }
    DistrictCode:
      name: code
      in: path
      required: true
      schema: { $ref: "#/components/schemas/DistrictCode" }
    MetricKey:
      name: key
      in: path
      required: true
      schema: { type: string, examples: [sale_price_median_m2] }
  headers:
    ETag:
      description: Derived from `data_version`. Send it back in `If-None-Match` to get 304.
      schema: { type: string }
    CacheControl:
      description: '`public, max-age=300`; shorter for weekly-refresh metrics.'
      schema: { type: string }
    DataWarning:
      description: Present when the data is stale or the last audit failed. The reasons are in `/meta` under `stale`.
      schema: { type: string }
  responses:
    Problem:
      description: Error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Lang: { type: string, enum: [en, pl] }
    DistrictCode:
      type: string
      description: Stable slug of the district, for example `srodmiescie`.
      pattern: "^[a-z0-9-]+$"
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    Source:
      type: object
      required: [name, licence, attribution]
      properties:
        name: { type: string }
        url: { type: string, format: uri }
        licence: { type: string }
        attribution: { type: string, description: 'The credit line the source requires; show it beside the value or in the footer.' }
    Meta:
      type: object
      required: [city, default_lang, languages, district_count, data_version, stale, sources]
      properties:
        city: { type: string, examples: [warsaw] }
        city_name: { type: string, examples: [Warszawa] }
        default_lang: { $ref: "#/components/schemas/Lang" }
        languages: { type: array, items: { $ref: "#/components/schemas/Lang" } }
        district_count: { type: integer }
        data_version: { type: string, description: 'Hash of the latest fetch times and report versions; also the ETag source.' }
        latest_ingestion_run: { type: [string, "null"], format: date-time }
        stale:
          type: object
          required: [is_stale, reasons]
          properties:
            is_stale: { type: boolean }
            reasons:
              type: array
              items: { type: string }
              description: For example a failed audit, reports older than their data, or a weekly source that did not refresh.
        sources:
          type: array
          description: One entry per source in use, with the newest as-of date, for the footer.
          items:
            allOf:
              - { $ref: "#/components/schemas/Source" }
              - type: object
                properties:
                  as_of: { type: string, format: date }
                  metric_keys: { type: array, items: { type: string } }
        score_note: { type: string }
...
    Highlight:
      type: object
      required: [key, display, data_kind]
      properties:
        key: { type: string }
        display: { type: string }
        data_kind: { $ref: "#/components/schemas/DataKind" }
    DistrictListItem:
      type: object
      required: [code, name, area_km2, livability_score, highlights]
      properties:
        code: { $ref: "#/components/schemas/DistrictCode" }
        name: { type: string, description: 'A proper noun; the same in both languages.' }
        area_km2: { type: number }
        livability_score: { type: [number, "null"], minimum: 0, maximum: 100 }
        highlights:
          type: array
          description: Sale price and rent per m2, so the map list loads fast.
          items: { $ref: "#/components/schemas/Highlight" }
    DistrictDetail:
      type: object
      required: [code, name, area_km2, lang, livability_score, categories]
      properties:
        code: { $ref: "#/components/schemas/DistrictCode" }
        name: { type: string }
        area_km2: { type: number }
        lang: { $ref: "#/components/schemas/Lang" }
        livability_score: { type: [number, "null"], minimum: 0, maximum: 100 }
        score_note: { type: string }
        categories:
          type: array
          items: { $ref: "#/components/schemas/CategoryGroup" }
        yield_gross:
          description: Gross rental yield (rent x 12 / price) from the district medians; null if either input is missing.
          oneOf: [{ $ref: "#/components/schemas/MetricValue" }, { type: "null" }]
        payback_years:
          oneOf: [{ $ref: "#/components/schemas/MetricValue" }, { type: "null" }]
    BoundaryCollection:
      type: object
      required: [type, features]
      properties:
        type: { const: FeatureCollection }
        features:
          type: array
          items:
            type: object
            required: [type, geometry, properties]
            properties:
              type: { const: Feature }
              geometry: { type: object, description: 'A GeoJSON Polygon or MultiPolygon, simplified.' }
              properties:
                type: object
                required: [code, name]
                properties:
                  code: { $ref: "#/components/schemas/DistrictCode" }
                  name: { type: string }
    Report:
      type: object
      required: [district, lang, lang_fallback, livability_score, body, generated_at]
      properties:
        district: { $ref: "#/components/schemas/DistrictCode" }
        lang: { $ref: "#/components/schemas/Lang" }
        lang_fallback: { type: boolean, description: 'True if the requested language was missing and English is returned.' }
        livability_score: { type: [number, "null"] }
        body: { type: string }
        model: { type: string }
        generated_at: { type: string, format: date-time }
        estimated_or_proxy_values_flagged: { type: boolean }
...
    Series:
      type: object
      required: [district, key, lang, min_obs, points]
      properties:
        district: { $ref: "#/components/schemas/DistrictCode" }
        key: { type: string }
        label: { type: string }
        lang: { $ref: "#/components/schemas/Lang" }
        unit: { type: string }
        data_kind: { $ref: "#/components/schemas/DataKind" }
        min_obs: { type: integer, description: 'Quarters with fewer observations are marked low confidence. Proposal 30.' }
        source: { $ref: "#/components/schemas/Source" }
        method: { type: string }
        caveat: { type: [string, "null"] }
        points:
          type: array
          items:
            type: object
            required: [period_start, period_end, value, display, n_obs, low_confidence]
            properties:
              period_start: { type: string, format: date }
              period_end: { type: string, format: date }
              value: { type: number }
              display: { type: string }
              n_obs: { type: integer }
              low_confidence: { type: boolean }
    Outlook:
      type: object
      required: [district, lang, momentum, city_history, scenario, method, caveat]
      properties:
        district: { $ref: "#/components/schemas/DistrictCode" }
        lang: { $ref: "#/components/schemas/Lang" }
        data_kind: { const: observed }
        momentum:
          type: object
          required: [growth_12m, growth_since_start]
          properties:
            growth_12m: { oneOf: [{ $ref: "#/components/schemas/Growth" }, { type: "null" }], description: 'Null if the history has no quarter one year before the newest one.' }
            growth_since_start: { oneOf: [{ $ref: "#/components/schemas/Growth" }, { type: "null" }] }
        city_history:
          type: object
          description: How much prices in the whole city changed over past windows. A record of the past, not a forecast.
          required: [available]
          properties:
            available: { type: boolean }
            period_start: { type: string, format: date }
            period_end: { type: string, format: date }
            windows:
              type: array
              items:
                type: object
                required: [quarters, label, low, median, high, low_display, median_display, high_display, n_windows]
                properties:
                  quarters: { type: integer, description: 'Length of the window: 4 or 8 quarters.' }
                  label: { type: string }
                  low: { type: number, description: '10th percentile of past price changes, as a fraction.' }
                  median: { type: number }
                  high: { type: number, description: '90th percentile.' }
                  low_display: { type: string }
                  median_display: { type: string }
                  high_display: { type: string }
                  n_windows: { type: integer, description: 'Overlapping windows behind the figures; far fewer are independent.' }
            source: { $ref: "#/components/schemas/Source" }
            method: { type: string }
        scenario:
          type: object
          required: [published, reason]
          properties:
            published: { type: boolean, description: 'False while no tested method passes the backtest.' }
            reason: { type: string }
            ranges: { type: array, items: {}, description: 'Reserved for a passing method; empty while `published` is false.' }
            backtest:
              type: object
              properties:
                period_start: { type: string, format: date }
                period_end: { type: string, format: date }
                cities: { type: integer }
                results:
                  type: array
                  items:
                    type: object
                    required: [quarters, method, mae_method, mae_no_change, mae_last_year_continues, coverage_80, passes]
                    properties:
                      quarters: { type: integer }
                      method: { type: string }
                      origins: { type: integer }
                      mae_method: { type: number, description: 'Median absolute error of the method, in log growth.' }
                      mae_no_change: { type: number }
                      mae_last_year_continues: { type: number }
                      coverage_80: { type: number, description: 'Share of outcomes inside the 10th-90th percentile range; 0.8 is ideal.' }
                      passes: { type: boolean }
        method: { type: string }
        caveat: { type: string }
...
    RentVsBuy:
      type: object
      required: [district, lang, area_m2, price, monthly_rent, yield_gross, payback_years, data_kind, caveat]
      properties:
        district: { $ref: "#/components/schemas/DistrictCode" }
        lang: { $ref: "#/components/schemas/Lang" }
        area_m2: { type: number }
        price: { $ref: "#/components/schemas/Amount" }
        monthly_rent: { $ref: "#/components/schemas/Amount" }
        yield_gross: { type: number, description: 'A fraction, for example 0.055 for 5.5 percent.' }
        yield_display: { type: string }
        payback_years: { type: number }
        payback_display: { type: string }
        data_kind: { const: estimated }
        based_on: { type: array, items: { $ref: "#/components/schemas/Highlight" }, description: 'The stored medians used, with counts in `n_obs` of each.' }
        caveat: { type: string, description: 'Asking rent over transaction price; no mortgage, tax, costs or price change.' }
...
    Commute:
      type: object
      required: [from, lang, method, as_of, destinations]
      properties:
        from: { $ref: "#/components/schemas/DistrictCode" }
        lang: { $ref: "#/components/schemas/Lang" }
        data_kind: { const: estimated }
        method: { type: string }
        as_of: { type: string, format: date }
        caveat: { type: string }
        destinations:
          type: array
          items:
            type: object
            required: [code, minutes]
            properties:
              code: { $ref: "#/components/schemas/DistrictCode" }
              minutes: { type: [number, "null"], description: 'Null if no public-transport connection was found within the search limit.' }
    Persona:
      type: object
      required: [key, label, description, weights]
      properties:
        key: { type: string, examples: [student] }
        label: { type: string }
        description: { type: string }
        weights: { $ref: "#/components/schemas/Weights" }
...
    RecommendRequest:
      type: object
      properties:
        weights: { $ref: "#/components/schemas/Weights" }
        lang: { $ref: "#/components/schemas/Lang" }
...